split usage/help message into separate lines
Instead of being on a single long line in the source code, the usage
message is now on separate lines corresponding to the lines as they would
appear in the output.  This makes the usage message easier for a developer
to read and edit.

This commit makes no functional changes; it is purely asthetic.